Do not exit when hitting non-initialized submodule. Print them instead.
Sometimes we do not want to include some modules, but also are not
interested in setting .gitattributes

For example, Project A depends on Project B, and B contains a library
for A and a visualizer. This visualizer will further depends on a GUI
Library C, but both the visualizer and the Library C are not used by A.
Unfortunately, B has already set .gitattributes for C.

Now, consider we want to package A along with its dependencies but
without C. Why we are forced to modify .gitattributes in Project B, and
damaging the integrity of Project B to achieve this?

This patch allows a non-intrusive way to exclude submodule. Just do not
`git init` them in the beginning, build the project for test, and
git-archive-all to get the re-distributable source package. Case closed.